Rename registry.ts to blockRegistry.ts
