+++ src/core/blockly/setupToolbox.ts
@@ -1,4 +1,4 @@
-import {registerBlockSpecs} from "./specs/registry.ts";
+import {registerBlockSpecs} from "./specs/blockRegistry.ts";
 import {applyCategoryColours} from "./colours.ts";
 
 export function setupToolbox() {
